Move the analysis to master
Currently the wrapper calls Ninka
