+++ qmstr/qmstr-analyze/analyzer.go
@@ -1,12 +1,10 @@
 package analyze
 
 import (
-	"encoding/csv"
 	"io"
 	"io/ioutil"
 	"log"
 	"os"
-	"os/exec"
 )
 
 var (
@@ -22,34 +20,4 @@ func initLogging(debug bool) {
 		infoWriter = ioutil.Discard
 	}
 	Logger = log.New(infoWriter, "", log.Ldate|log.Ltime)
-}
-
-func analyzeSourceFile(sourcefile string) []string {
-	licenses := []string{}
-	cmd := exec.Command("ninka", "-i", sourcefile)
-	err := cmd.Start()
-	checkErr(err)
-	if err := cmd.Wait(); err != nil {
-		log.Fatalf("License analysis failed for %s", sourcefile)
-	}
-
-	licenseFile, err := os.Open(sourcefile + ".license")
-	checkErr(err)
-	r := csv.NewReader(licenseFile)
-	r.Comma = ';'
-	records, err := r.ReadAll()
-	checkErr(err)
-
-	for _, fields := range records {
-		if len(fields) > 0 {
-			licenses = append(licenses, fields[0])
-		}
-	}
-	return licenses
-}
-
-func checkErr(err error) {
-	if err != nil {
-		log.Fatal(err)
-	}
 }
